refactor(versions): move fetch/format utilities to utils/
Per AGENTS.md: utility functions belong in utils/. Both fetchPackageInfo
and formatPublishedDate (and their tests) are generic — nothing about
them is specific to the Versions block — so they move out of
components/blocks/Versions/ into utils/ alongside the rest of the
repo's utilities.
