+++ content/main/versions.json
@@ -8,6 +8,26 @@
     {
       "title": "TinaCMS Package Versions",
       "description": "The currently published version of every package in the TinaCMS ecosystem - pulled straight from the npm registry on page load.",
+      "buttons": [
+        {
+          "label": "TinaCMS Changeset",
+          "icon": true,
+          "shape": "square",
+          "variant": "orange",
+          "size": "medium",
+          "url": "/whats-new/tinacms",
+          "_template": "actions"
+        },
+        {
+          "label": "TinaCloud Changeset",
+          "icon": true,
+          "shape": "square",
+          "variant": "orangeWithBorder",
+          "size": "medium",
+          "url": "/whats-new/tinacloud",
+          "_template": "actions"
+        }
+      ],
       "_template": "versions"
     }
   ]
