+++ utils/fetchPackageInfo.test.ts
